+++ commands/generic/bot.js
@@ -25,9 +25,6 @@ class Bot extends Command {
     }
 
     async run(client, message) {
-        if (client.bot.uptime < 60000) {
-            return message.channel.createMessage(':x: I am still booting up ! Please try again in a minute');
-        }
         return this.sendStats(client, message);
     }
 
@@ -91,9 +88,9 @@ class Bot extends Command {
             inline: true
         });
         embedFields.push({
-            name: "Developer",
-            value: "ParadoxOrigins#5451",
-            inline: true
+            name: "Developers",
+            value: "**Lead Developer**: ParadoxOrigins#5451\n**Co-Developers**: Ota#1354, Niputi#2490\n**Contributors**: InternalCosmos#2000, LevitatingBusinessMan#0504",
+            inline: false
         });
         embedFields.push({
             name: "Created the",
@@ -125,22 +122,31 @@ class Bot extends Command {
             value: '[Patreon](https://www.patreon.com/paradoxorigins)',
             inline: false
         });
-        embedFields.push({
-            name: `Shard`,
-            value: (() => {
-                let shardCount = 0;
-                for (const cluster of client.stats.clusters) {
-                    shardCount = shardCount + cluster.shards;
-                }
-                return `${message.channel.guild.shard.id}/${shardCount}`;
-            })(),
-            inline: true
-        });
+        if (client.stats) {
+            embedFields.push({
+                name: `Shard`,
+                value: (() => {
+                    let shardCount = 0;
+                    for (const cluster of client.stats.clusters) {
+                        shardCount = shardCount + cluster.shards;
+                    }
+                    return `${message.channel.guild.shard.id}/${shardCount}`;
+                })(),
+                inline: client.redis ? false : true
+            });
+        }
         embedFields.push({
             name: 'Database status',
-            value: client.database && client.database.healthy ? ':white_check_mark: Online' : ':x: Offline',
+            value: `${client.database && client.database.healthy ? ':white_check_mark: Online' : ':x: Offline'}\n[More info](https://github.com/ParadoxalCorp/felix-production/blob/master/usage.md#rethinkdb)`,
             inline: true
         });
+        if (client.redis) {
+            embedFields.push({
+                name: 'Redis status',
+                value: `${client.redis.healthy ? ':white_check_mark: Online' : ':x: Offline'}\n[More info](https://github.com/ParadoxalCorp/felix-production/blob/master/usage.md#redis)`,
+                inline: true
+            });
+        }
         return embedFields;
     }
 }